AI-Employee/requirements.txt
# =============================================
# AI Employee — Gold Tier Dependencies
# =============================================
# Install: pip install -r requirements.txt
# =============================================

# --- Core ---
python-dotenv>=1.0.0        # .env file loading
watchdog>=4.0.0              # File system monitoring

# --- AI / LLM ---
anthropic>=0.42.0            # Claude API (Task Intelligence Engine)

# --- Google Gmail API ---
google-api-python-client>=2.100.0   # Gmail API client
google-auth-httplib2>=0.2.0         # HTTP/2 auth transport
google-auth-oauthlib>=1.2.0         # OAuth2 flow for Gmail

# --- Web Dashboard ---
fastapi>=0.115.0             # Web dashboard framework
uvicorn[standard]>=0.32.0    # ASGI server for FastAPI
jinja2>=3.1.0                # HTML template rendering

# --- Odoo MCP Server ---
mcp>=1.0.0                       # Model Context Protocol SDK
requests>=2.31.0                 # HTTP client for Odoo JSON-RPC

# --- Twitter (X) API ---
requests-oauthlib>=1.3.0         # OAuth 1.0a for Twitter API v2 writes

# --- Already in stdlib (no install needed) ---
# smtplib                    # Gmail SMTP (legacy fallback)
# http.server                # Legacy dashboard server
# json, logging, pathlib     # Core utilities
